requirements.txt
numpy~=2.3.5
shapely~=2.1.2
scipy~=1.16.3
PyYAML~=6.0.3
trimesh~=4.10.1
pyvista~=0.46.4
pytest~=9.0.2